+++ lib/rdf/blank_node_generator/algorithm.ex
@@ -5,10 +5,10 @@ defmodule RDF.BlankNode.Generator.Algorithm do
   The `RDF.BlankNode.Generator` executes such an algorithm and holds its state.
   """
 
-  @doc """
-  Returns the initial state of the algorithm.
-  """
-  @callback init(opts :: map) :: map
+  @type type :: module
+  @type t :: struct
+
+  alias RDF.BlankNode
 
   @doc """
   Generates a blank node.
@@ -16,16 +16,16 @@ defmodule RDF.BlankNode.Generator.Algorithm do
   An implementation should compute a blank node from the given state and return
   a tuple consisting of the generated blank node and the new state.
   """
-  @callback generate(state :: map) :: {RDF.BlankNode.t(), map}
+  @callback generate(t()) :: {BlankNode.t(), t()}
 
   @doc """
   Generates a blank node for a given string.
 
   Every call with the same string must return the same blank node.
 
-  An implementation should compute a blank node for the given string from the
+  An implementation should compute a blank node for the given value from the
   given state and return a tuple consisting of the generated blank node and the
   new state.
   """
-  @callback generate_for(value :: any, state :: map) :: {RDF.BlankNode.t(), map}
+  @callback generate_for(t(), value :: any) :: {BlankNode.t(), t()}
 end