+++ lib/rdf/blank_node_generator/random.ex
@@ -1,62 +1,54 @@
 defmodule RDF.BlankNode.Generator.Random do
   @moduledoc """
   An implementation of a `RDF.BlankNode.Generator.Algorithm` which returns `RDF.BlankNode`s with random identifiers.
-
-  The following options are supported when starting a `RDF.BlankNode.Generator`
-  with this algorithm:
-
-  - `prefix`: a string prepended to the generated blank node identifier
-
   """
 
   @behaviour RDF.BlankNode.Generator.Algorithm
 
-  alias RDF.BlankNode
+  defstruct prefix: "b", map: %{}
 
-  @type state :: %{
-          optional(:prefix) => String.t(),
+  @type t :: %__MODULE__{
+          prefix: String.t(),
           map: map
         }
 
-  @impl BlankNode.Generator.Algorithm
-  def init(%{prefix: prefix} = opts) do
-    opts
-    |> Map.delete(:prefix)
-    |> init()
-    |> Map.put(:prefix, prefix)
+  alias RDF.BlankNode
+
+  @doc """
+  Creates a struct with the state of the algorithm.
+
+  ## Options
+
+  - `prefix`: a string prepended to the generated blank node identifier
+  """
+  def new(attrs \\ []) do
+    struct(__MODULE__, attrs)
   end
 
   @impl BlankNode.Generator.Algorithm
-  def init(_opts) do
-    %{
-      map: %{}
-    }
+  def generate(%__MODULE__{} = state) do
+    {bnode(state, number(state)), state}
   end
 
   @impl BlankNode.Generator.Algorithm
-  def generate(state) do
-    {state |> number() |> bnode(state), state}
-  end
-
-  @impl BlankNode.Generator.Algorithm
-  def generate_for(value, %{map: map} = state) do
+  def generate_for(%__MODULE__{map: map} = state, value) do
     case Map.get(map, value) do
       nil ->
         random = number(state)
-        {bnode(random, state), %{state | map: Map.put(map, value, random)}}
+        {bnode(state, random), %{state | map: Map.put(map, value, random)}}
 
       previous ->
-        {bnode(previous, state), state}
+        {bnode(state, previous), state}
     end
   end
 
-  defp bnode(random, %{prefix: prefix}) do
-    BlankNode.new(prefix <> random)
-  end
-
-  defp bnode(random, _state) do
+  defp bnode(%__MODULE__{prefix: nil}, random) do
     BlankNode.new(random)
   end
 
-  defp number(_), do: :erlang.unique_integer([:positive]) |> to_string()
+  defp bnode(%__MODULE__{} = state, random) do
+    BlankNode.new(state.prefix <> random)
+  end
+
+  defp number(_), do: :erlang.unique_integer([:positive]) |> Integer.to_string()
 end